Mostly working - tpm2 key loading is still flaky.
This is a snapshot before I refactor the init - this attempt may work
but it's still too complicated, tpm2 has too many deps.

The original goal was for initrd to be very simple and clean - and only
load the root. Having the dm-verity squashfs ( can also use partitions
like chromeos ) allows secure load of the 'recovery/install/init' rootfs
and that has full firmware/modules/tools - so we can pivot twice.