Finally got NVidia to work reliably. In the process decided to simplify again - while initializing NVidia using Debian kernel and Alpine rootfs works, it is too complicated and fragile and to use it the versions and 'user' space need to be kept in sync, etc.
So back to original idea of Debian based rootfs, with minimal systemd use and Alpine as a sidecar
handling networking and updates. Instead of switching root 2x, it goes directly from Initrd to the
read-only signed debian.

Keeping the Alpine-based rootfs around for now, but may move it to an example dir